Fix error with python3 and openbsd httpd
Without strict_slashes=False on root url, flask do 301 redirection to
add the slash but openbsd's httpd remove the slash. Or something like
that. Anyway, the result is a redirection loop without that change.

With python3, httplib2 return content as byte but json take a string. So
without a decode, a type error was raised on log in.